source/requirements.txt
flask
flask-jwt-extended
flask-pymongo
flask-mail
werkzeug
firebase-admin
python-dotenv
gunicorn
